Remove upper bound limit from our dependencies
All that was needed to support Rails 7 in this gem was to update the
gemspec. This means that even in the most likely release to break
anything, a major release, we didn't needed to worry about breaking this
gem.

This upper bound is just creating unnecessary work for the
maintainers of this gem, which would need to release it without any code
change, and the users, which need to open issues asking for new releases.

Let's allow our users to try any version of our dependencies and report
real issues if they find them, and save everyone's time just removing
the upper bound.